Descriptor formatting adjustments

+++ ORB_SLAM3/include/ORBmatcher.h
@@ -39,10 +39,10 @@ public:
     ORBmatcher(float nnratio=0.6, bool checkOri=true);
 
     // Computes the Hamming distance between two ORB descriptors
-    static int DescriptorDistance(const cv::Mat &a, const cv::Mat &b);
+    static float DescriptorDistance(const cv::Mat &a, const cv::Mat &b);
 
     // Computes L2 Norm between two SuperPoint Descriptors
-    static int DescriptorDistanceL2(const cv::Mat &a, const cv::Mat &b);
+    static float DescriptorDistanceL2(const cv::Mat &a, const cv::Mat &b);
 
     // Search matches between Frame keypoints and projected MapPoints. Returns number of matches
     // Used to track the local map (Tracking)
@@ -92,8 +92,8 @@ public:
 
 public:
 
-    static const int TH_LOW;
-    static const int TH_HIGH;
+    static const float TH_LOW;
+    static const float TH_HIGH;
     static const int HISTO_LENGTH;
